Reverted renaming changes
