+++ lib/networkx.rb
@@ -1,4 +1,4 @@
 require 'set'
 
 require 'networkx/version'
-require 'networkx/classes/graph'
+require 'networkx/graph'
