+++ lib/networkx/graph.rb
